Rename getExistenceQuery -> getExistenceCheckQuery
