+++ src/neo4j/cypher-queries/validation/index.js
@@ -1,12 +1,12 @@
 import getDuplicateRecordCheckQuery from './duplicate-record-check';
-import getExistenceQuery from './existence';
+import getExistenceCheckQuery from './existence-check';
 import getSubMaterialChecksQuery from './sub-material-checks';
 import getSubProductionChecksQuery from './sub-production-checks';
 import getSubVenueChecksQuery from './sub-venue-checks';
 
 export {
 	getDuplicateRecordCheckQuery,
-	getExistenceQuery,
+	getExistenceCheckQuery,
 	getSubMaterialChecksQuery,
 	getSubProductionChecksQuery,
 	getSubVenueChecksQuery
